Rename rich text code extension module
